CLAI2009

Description

Código correspondiente al artículo "Sistema certificado de decisión proposicional basado en polinomios" presentado en el CLAI2009 (Workshop on Computational Logic and Artificial Intelligence).

Lógica

Sintaxis de fórmulas proposicionales

type SimboloProposicional = String

Los símbolos proposicionales se representan por cadenas.

data FProp

FProp es el tipo de las fórmulas proposicionales definidas por

T y F son fórmulas

Si A es una fórmula, también lo es ¬A.

Si A y B son fórmulas, entonces (A ∧ B), (A ∨ B), (A → B) y (A ↔ B) también lo son.

Constructors

T

F

Atom SimboloProposicional

Neg FProp

Conj FProp FProp

Disj FProp FProp

Impl FProp FProp

Equi FProp FProp

Ejemplos de fórmulas atómicas

p :: FProp

Ejemplo de fórmulas.

q :: FProp

Ejemplo de fórmulas.

r :: FProp

Ejemplo de fórmulas.

Conectivas lógicas

no :: FProp -> FProp

(no f) es la negación de f

(∨) :: FProp -> FProp -> FProp infixr 5

(f ∨ g) es la disyunción de f y g.

(∧) :: FProp -> FProp -> FProp infixr 4

(f ∧ g) es la conjunción de f y g

(→) :: FProp -> FProp -> FProp infixr 3

(f → g) es la implicación de f a g

(↔) :: FProp -> FProp -> FProp infixr 2

(f ↔ g) es la equivalencia entre f y g

Semántica de la lógica proposicional

type Interpretacion = [FProp]

Las interpretaciones son listas de fórmulas atómicas. Las fórmulas de las interpretaciones se suponen verdaderas y las restantes fórmulas atómicas se suponen falsas.

significado :: FProp -> Interpretacion -> Bool

(significado f i) es el significado de la fórmula f en la interprestación i. Por ejemplo,

>>> significado ((p ∨ q) ∧ ((no q) ∨ r)) [r]
False
>>> significado ((p ∨ q) ∧ ((no q) ∨ r)) [p,r]
True

Modelos de una fórmula

simbolosPropForm :: FProp -> [FProp]

(simbolosPropForm f) es el conjunto formado por todos los símbolos proposicionales que aparecen en f. Por ejemplo,

>>> simbolosPropForm (p ∧ q → p)
[p,q]

interpretacionesForm :: FProp -> [Interpretacion]

(interpretacionesForm f) es la lista de todas las interpretaciones de la fórmula f. Por ejemplo,

>>> interpretacionesForm (p ∧ q → p)
[[],[p],[q],[p,q]]

esModeloFormula :: Interpretacion -> FProp -> Bool

(esModeloFormula i f) se verifica si la interpretación i es un modelo de la fórmula f. Por ejemplo,

>>> esModeloFormula [r]   ((p ∨ q) ∧ ((no q) ∨ r))
False
>>> esModeloFormula [p,r] ((p ∨ q) ∧ ((no q) ∨ r))
True

modelosFormula :: FProp -> [Interpretacion]

(modelosFormula f) es la lista de todas las interpretaciones de la fórmula f que son modelo de F. Por ejemplo,

>>> modelosFormula ((p ∨ q) ∧ ((no q) ∨ r))
[[p],[p,r],[q,r],[p,q,r]]

Fórmulas válidas, satisfacibles e insatisfacibles

esValida :: FProp -> Bool

(esValida f) se verifica si la fórmula f es válida. Por ejemplo,

>>> esValida (p → p)
True
>>> esValida (p → q)
False
>>> esValida ((p → q) ∨ (q → p))
True

prop_esValida :: FProp -> Bool

(prop_esValida f) se verifica si las siguiente condiciones son equivalentes:

+ f es válida + f es consecuencia del conjunto vacío.

prop_esValida

esInsatisfacible :: FProp -> Bool

esSatisfacible :: FProp -> Bool

Interpretaciones de un conjunto de fórmulas

unionGeneral :: Eq a => [[a]] -> [a]

simbolosPropConj :: [FProp] -> [FProp]

interpretacionesConjunto :: [FProp] -> [Interpretacion]

Modelos de conjuntos de fórmulas

esModeloConjunto :: Interpretacion -> [FProp] -> Bool

modelosConjunto :: [FProp] -> [Interpretacion]

Conjuntos consistentes e inconsistentes de fórmulas

esConsistente :: [FProp] -> Bool

esInconsistente :: [FProp] -> Bool

Consecuencia lógica

esConsecuencia :: [FProp] -> FProp -> Bool

prop_esConsecuencia :: [FProp] -> FProp -> Bool

Equivalencia de fórmulas

equivalentes :: FProp -> FProp -> Bool

Polinomios

Variables

caracteres :: Gen Char

type Variable = String

variables :: Gen String

variables' :: Gen String

Monomios

data Monomio

Constructors

M [Variable]

mUno :: Monomio

esMonomio :: Monomio -> Bool

monomiosN :: Int -> Gen Monomio

monomios :: Gen Monomio

monomios' :: Gen Monomio

prop_MonomiosGeneraMonomios :: Monomio -> Bool

Polinomios

data Polinomio

Constructors

P [Monomio]

cero :: Polinomio

uno :: Polinomio

esPolinomio :: Polinomio -> Bool

polinomiosN :: Int -> Gen Polinomio

polinomios :: Gen Polinomio

polinomios' :: Gen Polinomio

prop_PolinomiosGeneraPolinomios :: Polinomio -> Bool

Suma de polinomios

suma :: Polinomio -> Polinomio -> Polinomio

sumaAux :: [Monomio] -> [Monomio] -> [Monomio]

prop_suma_bien_definida :: Polinomio -> Polinomio -> Bool

prop_suma_conmutativa :: Polinomio -> Polinomio -> Bool

prop_suma_asociativa :: Polinomio -> Polinomio -> Polinomio -> Bool

prop_suma_neutro :: Polinomio -> Bool

prop_suma_simetrico :: Polinomio -> Bool

prop_distributiva :: Polinomio -> Polinomio -> Polinomio -> Bool

Productos de polinomios

productoMM :: Monomio -> Monomio -> Monomio

productoMP :: Monomio -> Polinomio -> Polinomio

producto :: Polinomio -> Polinomio -> Polinomio

prop_prod_bien_definido :: Polinomio -> Polinomio -> Bool

prop_prod_conmutativa :: Polinomio -> Polinomio -> Bool

Derivada de polinomios

deriv :: Polinomio -> Variable -> Polinomio

prop_deriv_bien_definida :: Polinomio -> Variable -> Bool

prop_deriv_deriv :: Polinomio -> Variable -> Bool

prop_deriv_suma :: Polinomio -> Polinomio -> Variable -> Bool

prop_deriv_prod :: Polinomio -> Polinomio -> Variable -> Bool

Transformación de proposiciones en polinomios

tr :: FProp -> Polinomio

pro_tr_bien_definida :: FProp -> Bool

Transformación de polinomios en proposiciones

theta :: Polinomio -> FProp

thetaAux :: [Monomio] -> FProp

theta2 :: Monomio -> FProp

theta2Aux :: [SimboloProposicional] -> FProp

prop_theta_tr :: FProp -> Bool

prop_tr_theta :: Polinomio -> Bool

Derivación de fórmulas proposicionales

derivP :: FProp -> SimboloProposicional -> FProp

sustituir :: FProp -> SimboloProposicional -> FProp -> FProp

prop_derivP_sustituir :: FProp -> Bool

variablesProp :: FProp -> [SimboloProposicional]

La regla de independencia (regla delta)

indep :: Polinomio -> Variable -> Polinomio

delta :: Polinomio -> Polinomio -> Variable -> Polinomio

delta' :: Polinomio -> Polinomio -> Variable -> Polinomio

prop_equiv_delta_delta' :: Polinomio -> Polinomio -> Bool

variablesMon :: Monomio -> [Variable]

variablesPol :: Polinomio -> [Variable]

deltaP :: FProp -> FProp -> SimboloProposicional -> FProp

Adecuación y completitud de la regla de independencia

prop_adecuacion_deltaP :: FProp -> FProp -> Bool

pares :: [a] -> [(a, a)]

derivadas :: [Polinomio] -> Variable -> [Polinomio]

derivadasP :: [FProp] -> SimboloProposicional -> [FProp]

deltaRefutable :: [Polinomio] -> Bool

deltaRefutableP' :: [FProp] -> Bool

prop_def_alt_deltaRefutableP :: [FProp] -> Bool

prop_adecuacion_completitud_deltaP :: [FProp] -> Bool

deltaDemostrable :: [FProp] -> FProp -> Bool

prop_adecuacion_completitud_delta_2 :: [FProp] -> FProp -> Bool

deltaTeorema :: FProp -> Bool

prop_adecuacion_completitud_delta_3 :: FProp -> Bool

deltaRefutableP :: [FProp] -> Bool

Delta refutabilidad usando soporte

deltaRefutableSop :: [Polinomio] -> Bool

deltaRefutableSop' :: [Polinomio] -> [Polinomio] -> Bool

derivadasPolConjunto :: Polinomio -> [Polinomio] -> [Polinomio]

derivadasPolPol :: Polinomio -> Polinomio -> [Polinomio]

prop_deltaRefutableSop :: [Polinomio] -> Bool

deltaRefutablePSop :: [FProp] -> Bool

prop_adecuacion_completitud_deltaPSop :: [FProp] -> Bool

Problema del palomar

palomar :: [FProp]

Bibliografía