DeltaDeduccion

Derivación de fórmulas proposicionales

derivP :: FProp -> SimboloProposicional -> FProp

(derivP f v) es la derivada de la fórmula proposicional f respecto de la variable v. Por ejemplo,

>>> derivP (p ∧ q → r) "p"
¬(q ↔ (q ∧ r))
>>> derivP (p ∧ q → r) "q"
¬(p ↔ (p ∧ r))
>>> derivP (p ∧ q → r) "r"
(p ∧ q)
>>> derivP (p ∧ q → p ∨ q) "p"
⊥

sustituir :: FProp -> SimboloProposicional -> FProp -> FProp

(sustituir f v g) es la fórmula obtenida sustituyendo en la fórmula f la variable v por la fórmula g. Por ejemplo,

>>> sustituir (p → q ∨ r) "q" (p ∧ q)
(p → ((p ∧ q) ∨ r))

variablesProp :: FProp -> [SimboloProposicional]

(variablesProp f) es la lista de las variables proposicionales de la fórmula f. Por ejemplo,

>>> variablesProp (p → q ∨ p)
["p","q"]

prop_derivP_sustituir :: FProp -> Bool

Comprueba que, para toda fórmula f y toda variable x de f, la derivada de f respecto de x es equivalente a la fórmula ¬(f[x/¬x] ↔ f)

>>> quickCheck prop_derivP_sustituir
+++ OK, passed 100 tests.

La regla de independencia (regla delta)

indep :: Polinomio -> Variable -> Polinomio

(indep p x) es el polinomio formado por los monomios de que no contienen la variable x. Por ejemplo,

>>> indep (P[M["x","y"],M["y","u"],M ["z"]]) "x"
y*u+z

delta :: Polinomio -> Polinomio -> Variable -> Polinomio

(delta a1 a2 v) es el polinomio obtenido aplicando la regla delta (o de independencia) a los polinomios a1 y a2 respecto de la variable v; es decir, el polinomio 1 + (1+a1*a2)*(1+a1*c2+a2*c1+c1*c2) donde ci es la derivada de ai respecto de v (para i = 1, 2). Por ejemplo,

>>> delta (P[M["x","y"],M["y","z"]]) (P[M["u","x"],M["y","z"]]) "x"
u*y+u*y*z+y*z

delta' :: Polinomio -> Polinomio -> Variable -> Polinomio

(delta' a1 a2 v) es el polinomio 1 + (1+b1*b2)*(1+(b1+c1)*(b2+c2)) donde ci es la derivada de ai respecto de v y bi es es el polinomio formado por los monomios de ai que no contienen la variable x (para i = 1, 2). Por ejemplo,

>>> delta' (P[M["x","y"],M["y","z"]]) (P[M["u","x"],M["y","z"]]) "x"
u*y+u*y*z+y*z

prop_equiv_delta_delta' :: Polinomio -> Polinomio -> Bool

Comprueba que las funciones delta y delta' son equivalentes.

>>> quickCheck prop_equiv_delta_delta'
+++ OK, passed 100 tests.

variablesMon :: Monomio -> [Variable]

(variablesMon m) es la lista de las variables del monomio m. Por ejemplo,

>>> variablesMon (M ["x","z"])
["x","z"]

variablesPol :: Polinomio -> [Variable]

(variablesPol p) es la lista de las variables del polinomio p. Por ejemplo,

>>> variablesPol (P[M["x","z"],M["y","z"]])
["x","z","y"]

deltaP :: FProp -> FProp -> SimboloProposicional -> FProp

(deltaP f1 f2 v) es la formula obtenida aplicando la regla delta (o de independencia) a las fórmulas f1 y f2 respecto de la variable v. Por ejemplo,

>>> deltaP (r → p ∨ q) (q → p ∧ q) "q"
¬(⊤ ↔ ¬((p ∧ r) ↔ r))

Adecuación y completitud de la regla de independencia

prop_adecuacion_deltaP :: FProp -> FProp -> Bool

Comprueba que la regla delta es adecuada: es decir que, para toda fórmulas f1 y f2 y toda variable x de ellas, la fórmula (deltaP f1 f2 x) es consecuencia de f1 y f2.

>>> quickCheck prop_adecuacion_deltaP
+++ OK, passed 100 tests.

pares :: [a] -> [(a, a)]

(pares xs) es la lista de los pares de elementos xs con el primero menor o igual que el segundo. Por ejemplo,

>>> pares [1..4]
[(1,1),(1,2),(1,3),(1,4),(2,2),(2,3),(2,4),(3,3),(3,4),(4,4)]

derivadas :: [Polinomio] -> Variable -> [Polinomio]

(derivadas ps x) es la lista de los polinomios obtenidos aplicando la regla delta a dos polinomios de ps respecto de la variable x. Por ejemplo,

>>> derivadas [P[M["x","y"],M["y","z"]],P[M["a","y"],M["y"]]] "y"
[x+z,a*x+a*z+x+z,1+a]

derivadasP :: [FProp] -> SimboloProposicional -> [FProp]

(derivadasP fs x) es la lista de las proposiciones obtenidas aplicando la regla deltaP a dos fórmulas de fs respecto de la variable x. Por ejemplo,

>>> derivadasP [p → q ∨ r, r → p] "q"
[¬(⊤ ↔ ¬((p ∧ r) ↔ r))]
>>> derivadasP [p → q ∨ r, r → q] "q"
[]

deltaRefutable :: [Polinomio] -> Bool

(deltaRefutable ps) se verifica si ps es refutable mediante la regla delta. Por ejemplo,

>>> deltaRefutable [P[mUno,M["p"],M["p","q"]],P[M["p"]],P[mUno,M["q"]]]
True
>>> deltaRefutable [P[mUno,M["p"],M["p","q"]],P[M["p"]],P[mUno,M["r"]]]
False

deltaRefutableP :: [FProp] -> Bool

(deltaRefutableP fs) se verifica si fs es refutable mediante la regla delta. Por ejemplo,

>>> deltaRefutableP [p → q, p, no q]
True
>>> deltaRefutableP [p → q, p, no r]
False

deltaRefutableP' :: [FProp] -> Bool

(deltaRefutableP' fs) se verifica si fs es refutable mediante la regla delta. Por ejemplo,

>>> deltaRefutableP' [p → q, p, no q]
True
>>> deltaRefutableP' [p → q, p, no r]
False

prop_def_alt_deltaRefutableP :: [FProp] -> Bool

Comprueba que las funciones deltaRefubleP y deltaRefutableP' son equivalentes.

>>> quickCheck prop_def_alt_deltaRefutableP
+++ OK, passed 100 tests.

prop_adecuacion_completitud_deltaP :: [FProp] -> Bool

Comprueba que la regla delta es adecuada y completa; es decir, para todo conjunto de fórmulas fs, fs es inconsistente si y sólo si es delta refutable.

>>> quickCheck prop_adecuacion_completitud_deltaP
+++ OK, passed 100 tests.

deltaDemostrable :: [FProp] -> FProp -> Bool

(deltaDemostrable fs g) se verifica si g es demostrable a partir de fs usando la regla delta. Por ejemplo,

>>> deltaDemostrable [p → q, q → r] (p → r)
True
>>> deltaDemostrable [p → q, q → r] (r → p)
False

prop_adecuacion_completitud_delta_2 :: [FProp] -> FProp -> Bool

Comprueba que la regla delta es adecuada y completa; es decir, para todo conjunto de fórmulas fs y toda fórmula g, g es consecuencia de fs si y sólo si es g es delta demostrable a partir de fs.

>>> quickCheck prop_adecuacion_completitud_delta_2
+++ OK, passed 100 tests.

deltaTeorema :: FProp -> Bool

(deltaTeorema f) se verifica si f es un teorema mediante la regla delta; es decir, si la negación de f es delta refutable. Por ejemplo,

>>> deltaTeorema ((p → q) ∨ (q → p))
True
>>> deltaTeorema ((p → q) ∨ (q → r))
True
>>> deltaTeorema ((p → q) ∨ (r → q))
False

prop_adecuacion_completitud_delta_3 :: FProp -> Bool

Comprueba que la regla delta es adecuada y completa; es decir, que para toda fórmula f, f es válida si y sólo si es un delta teorema.

>>> quickCheck prop_adecuacion_completitud_delta_3
+++ OK, passed 100 tests.

Bibliografía

Wikipedia: Zhegalkin polynomial.

G.A. Aranda, J. Borrego y M.M. Fernández (2009) Conservative retractions of propositional logic theories by means of boolean derivatives. Theoretical foundations.

M. Brickenstein y A. Dreyer: PolyBoRi: A Gröbner basis framework for Boolean polynomials.

M. Clegg, J. Edmonds y R. Impagliazzo (1996): Using the Groebner basis algorithm to find proofs of unsatisfiability.