Logica

Sintaxis de fórmulas proposicionales

type SimboloProposicional = String

Los símbolos proposicionales se representan por cadenas.

data FProp

FProp es el tipo de las fórmulas proposicionales definidas por

T y F son fórmulas

Si A es una fórmula, también lo es ¬A.

Si A y B son fórmulas, entonces (A ∧ B), (A ∨ B), (A → B) y (A ↔ B) también lo son.

Constructors

T

F

Atom SimboloProposicional

Neg FProp

Conj FProp FProp

Disj FProp FProp

Impl FProp FProp

Equi FProp FProp

Ejemplos de fórmulas atómicas

p :: FProp

Ejemplo de fórmulas.

q :: FProp

Ejemplo de fórmulas.

r :: FProp

Ejemplo de fórmulas.

Conectivas lógicas

no :: FProp -> FProp

(no f) es la negación de f

(∨) :: FProp -> FProp -> FProp infixr 5

(f ∨ g) es la disyunción de f y g.

(∧) :: FProp -> FProp -> FProp infixr 4

(f ∧ g) es la conjunción de f y g

(→) :: FProp -> FProp -> FProp infixr 3

(f → g) es la implicación de f a g

(↔) :: FProp -> FProp -> FProp infixr 2

(f ↔ g) es la equivalencia entre f y g

Semántica de la lógica proposicional

type Interpretacion = [FProp]

Las interpretaciones son listas de fórmulas atómicas. Las fórmulas de las interpretaciones se suponen verdaderas y las restantes fórmulas atómicas se suponen falsas.

significado :: FProp -> Interpretacion -> Bool

(significado f i) es el significado de la fórmula f en la interprestación i. Por ejemplo,

>>> significado ((p ∨ q) ∧ ((no q) ∨ r)) [r]
False
>>> significado ((p ∨ q) ∧ ((no q) ∨ r)) [p,r]
True

Modelos de una fórmula

simbolosPropForm :: FProp -> [FProp]

(simbolosPropForm f) es el conjunto formado por todos los símbolos proposicionales que aparecen en f. Por ejemplo,

>>> simbolosPropForm (p ∧ q → p)
[p,q]

interpretacionesForm :: FProp -> [Interpretacion]

(interpretacionesForm f) es la lista de todas las interpretaciones de la fórmula f. Por ejemplo,

>>> interpretacionesForm (p ∧ q → p)
[[],[p],[q],[p,q]]

esModeloFormula :: Interpretacion -> FProp -> Bool

(esModeloFormula i f) se verifica si la interpretación i es un modelo de la fórmula f. Por ejemplo,

>>> esModeloFormula [r]   ((p ∨ q) ∧ ((no q) ∨ r))
False
>>> esModeloFormula [p,r] ((p ∨ q) ∧ ((no q) ∨ r))
True

modelosFormula :: FProp -> [Interpretacion]

(modelosFormula f) es la lista de todas las interpretaciones de la fórmula f que son modelo de F. Por ejemplo,

>>> modelosFormula ((p ∨ q) ∧ ((no q) ∨ r))
[[p],[p,r],[q,r],[p,q,r]]

Fórmulas válidas, satisfacibles e insatisfacibles

esValida :: FProp -> Bool

(esValida f) se verifica si la fórmula f es válida. Por ejemplo,

>>> esValida (p → p)
True
>>> esValida (p → q)
False
>>> esValida ((p → q) ∨ (q → p))
True

prop_esValida :: FProp -> Bool

(prop_esValida f) se verifica si las siguiente condiciones son equivalentes:

f es válida

f es consecuencia del conjunto vacío.

>>> quickCheck prop_esValida
+++ OK, passed 100 tests.

esInsatisfacible :: FProp -> Bool

(esInsatisfacible f) se verifica si la fórmula f es insatisfacible. Por ejemplo,

>>> esInsatisfacible (p ∧ (no p))
True
>>> esInsatisfacible ((p → q) ∧ (q → r))
False

esSatisfacible :: FProp -> Bool

(esSatisfacible f) se verifica si f es satisfacible. Por ejemplo,

>>> esSatisfacible (p ∧ (no p))
False
>>> esSatisfacible ((p → q) ∧ (q → r))
True

Interpretaciones de un conjunto de fórmulas

unionGeneral :: Eq a => [[a]] -> [a]

(unionGeneral x) es la unión de los conjuntos de la lista de conjuntos x. Por ejemplo,

>>> unionGeneral []
[]
>>> unionGeneral [[1]]
[1]
>>> unionGeneral [[1],[1,2],[2,3]]
[1,2,3]

simbolosPropConj :: [FProp] -> [FProp]

(simbolosPropConj s) es el conjunto de los símbolos proposiciones de s. Por ejemplo,

>>> simbolosPropConj [p ∧ q → r, p → r]
[p,q,r]

interpretacionesConjunto :: [FProp] -> [Interpretacion]

(interpretacionesConjunto s) es la lista de las interpretaciones de s. Por ejemplo,

>>> interpretacionesConjunto [p → q, q → r]
[[],[p],[q],[p,q],[r],[p,r],[q,r],[p,q,r]]

Modelos de conjuntos de fórmulas

esModeloConjunto :: Interpretacion -> [FProp] -> Bool

(esModeloConjunto i s) se verifica si i es modelo de s. Por ejemplo,

>>> esModeloConjunto [p,r] [(p ∨ q) ∧ ((no q) ∨ r), q → r]
True
>>> esModeloConjunto [p,r] [(p ∨ q) ∧ ((no q) ∨ r), r → q]
False

modelosConjunto :: [FProp] -> [Interpretacion]

(modelosConjunto s) es la lista de modelos del conjunto s. Por ejemplo,

>>> modelosConjunto [(p ∨ q) ∧ ((no q) ∨ r), q → r]
[[p],[p,r],[q,r],[p,q,r]]
>>> modelosConjunto [(p ∨ q) ∧ ((no q) ∨ r), r → q]
[[p],[q,r],[p,q,r]]

Conjuntos consistentes e inconsistentes de fórmulas

esConsistente :: [FProp] -> Bool

(esConsistente s) se verifica si s es consistente. Por ejemplo,

>>> esConsistente [(p ∨ q) ∧ ((no q) ∨ r), p → r]
True
>>> esConsistente [(p ∨ q) ∧ ((no q) ∨ r), p → r, no r]
False

esInconsistente :: [FProp] -> Bool

(esInconsistente s) se verifica si s es inconsistente. Por ejemplo,

>>> esInconsistente [(p ∨ q) ∧ ((no q) ∨ r), p → r]
False
>>> esInconsistente [(p ∨ q) ∧ ((no q) ∨ r), p → r, no r]
True

Consecuencia lógica

esConsecuencia :: [FProp] -> FProp -> Bool

(esConsecuencia s f) se verifica si f es consecuencia de s. Por ejemplo,

>>> esConsecuencia [p → q, q → r] (p → r)
True
>>> esConsecuencia [p] (p ∧ q)
False

prop_esConsecuencia :: [FProp] -> FProp -> Bool

(prop_esConsecuencia s f) verifica que son equivalentes:

f es consecuencia de s

s ∪ {¬f} es inconsistente

>>> quickCheck prop_esConsecuencia
+++ OK, passed 100 tests.

Equivalencia de fórmulas

equivalentes :: FProp -> FProp -> Bool

(equivalentes f g) se verifica si las fórmulas f y g son equivalentes. Por ejemplo,

>>> equivalentes (p → q) (no p ∨ q)
True