PolinomiosF2

Variables

caracteres :: Gen Char

caracteres es un generador de caracteres de letras minúsculas. Por ejemplo,

   > sample caracteres
   x
   n
   r

type Variable = String

Las variables se representan por cadenas.

variables :: Gen String

variables es un generador de variables de longitud 1 ó 2. Por ejemplo,

   > sample variables
   "h"
   "yi"
   "m"

variables' :: Gen String

variables' es un generador de variables de longitud aleatoria, pero no cero. Por ejemplo,

   > sample variables'
   "o"
   "rte"
   "tmzeu"

Monomios

data Monomio

Los monomios son productos de variables distintas y se representan por listas ordenadas de variables distintas.

Constructors

M [Variable]

mUno :: Monomio

El monomio correspondiente a la lista vacía es el 1 (elemento neutro del producto).

>>> mUno
1

esMonomio :: Monomio -> Bool

La condición de que las variables sean distintas y ordenadas no se recoge en la definición del tipo de dato. Por ello se define el siguiente reconocedor de monomios. Por ejemplo,

>>> esMonomio (M ["x1","x3","z"])
True
>>> esMonomio (M ["x5","x3","z"])
False
>>> esMonomio (M ["x1","x1","z"])
False

monomiosN :: Int -> Gen Monomio

(monomiosN n) es un generador de monomios con el número de variables entre 0 y n. Por ejemplo,

   > sample (monomiosN 3)
   1
   s
   e*t
   hx*w*xn
   > sample (monomiosN 10)
   at*b*dy*fq*gv*mx*y*z
   a*cm*d*f*h*wf*z
   b*dw*wx*x*y*z

monomios :: Gen Monomio

monomios es un generador de monomios con el número de variables entre 0 y 3. Por ejemplo,

   > sample monomios
   nd*q
   e
   1

monomios' :: Gen Monomio

monomios' es un generador de monomios con un número aleatorio de variables. Por ejemplo,

   > sample monomios'
   1
   kl*o*u
   bm*d*k*mk

prop_MonomiosGeneraMonomios :: Monomio -> Bool

prop_MonomiosGeneraMonomios comprueba que el generador de monomios genera monomios. Por ejemplo,

>>> quickCheck prop_MonomiosGeneraMonomios
+++ OK, passed 100 tests.

Polinomios

data Polinomio

Los polinomios son sumas de monomios distintos y ordenados y se representan por listas ordenadas de monomios distintos,

Constructors

P [Monomio]

cero :: Polinomio

El polinomio correspondiente a la lista vacía es el 0 (elemento neutro de la suma). Por ejemplo,

>>> cero
0

uno :: Polinomio

El polinomio 1 es el polinomio cuyo único elemento es el monomio uno. Por ejemplo,

>>> uno
1

esPolinomio :: Polinomio -> Bool

La condición de que los monomios sean distintos y ordenados no se recoge en la definición del tipo de dato. Por ello se define el siguiente reconocedor de polinomios. Por ejemplo,

>>> esPolinomio (P [M ["a","b","c"], M ["x","y"]])
True
>>> esPolinomio (P [M ["x","y"], M ["a","c","d"]])
False
>>> esPolinomio (P [M ["x","y"], M ["a","d","c"]])
False
>>> esPolinomio (P [M ["x","y"], M ["a","a","c"]])
False
>>> esPolinomio (P [M ["x","y"], M ["x","y"], M ["a","c","d"]])
False
>>> esPolinomio (P [M ["a","a","c"], M ["x","y"]])
False

polinomiosN :: Int -> Gen Polinomio

(polinomiosN n) es un generador de polinomios con el número de monomios entre 0 y n. Por ejemplo,

   > sample (polinomiosN 3)
   0
   pp*sa
   gn*nf*zg
   > sample (polinomiosN 10)
   1+b*j+gw*w*zm+j*ox+l*q*qz+ly*p*r+m+q*zy
   iz
   1+d+dy+jd*l+lr+w

polinomios :: Gen Polinomio

polinomios es un generador de polinomios con el número de monomios entre 0 y 3. Por ejemplo,

   > sample monomios
   nd*q
   e
   1

polinomios' :: Gen Polinomio

polinomios' es un generador de polinomios con un número aleatorio de monomios. Por ejemplo,

   > sample polinomios'
   0
   1
   f*m*on*tr*ue*x
   ct*d*ds*gy*ps*s*y

prop_PolinomiosGeneraPolinomios :: Polinomio -> Bool

prop_PolinomiosGeneraPolinomios comprueba que el generador de polinomios genera polinomios. Por ejemplo,

>>> quickCheck prop_PolinomiosGeneraPolinomios
+++ OK, passed 100 tests.

Suma de polinomios

suma :: Polinomio -> Polinomio -> Polinomio

(suma p q) es la suma de los polinomios p y q. Por ejemplo,

>>> suma (P [M ["x"], M ["y"]]) (P [M ["y"], M ["z"]])
x+z
>>> suma (P [M ["x"], M ["y"]]) (P [M ["a"], M ["z"]])
a+x+y+z

sumaAux :: [Monomio] -> [Monomio] -> [Monomio]

(sumaAux xs ys) es la lista de las sumas de los monomios de xs e ys. Por ejemplo,

>>> sumaAux [M ["x"], M ["y"]] [M ["y"], M ["z"]]
[x,z]
>>> sumaAux [M ["x"], M ["y"]] [M ["a"], M ["z"]]
[a,x,y,z]

prop_suma_bien_definida :: Polinomio -> Polinomio -> Bool

Comprueba que la suma de polinomios está bien definida.

>>> quickCheck prop_suma_bien_definida
+++ OK, passed 100 tests.

prop_suma_conmutativa :: Polinomio -> Polinomio -> Bool

Comprueba que la suma de polinomios es conmutativa.

>>> quickCheck prop_suma_conmutativa
+++ OK, passed 100 tests.

prop_suma_asociativa :: Polinomio -> Polinomio -> Polinomio -> Bool

Comprueba que la suma de polinomios es conmutativa.

>>> quickCheck prop_suma_asociativa
+++ OK, passed 100 tests.

prop_suma_neutro :: Polinomio -> Bool

Comprueba que cero es el elemento neutro de la suma de polinomios.

>>> quickCheck prop_suma_neutro
+++ OK, passed 100 tests.

prop_suma_simetrico :: Polinomio -> Bool

Comprueba que cada elemento es su simétrico en la suma de polinomios.

>>> quickCheck prop_suma_simetrico
+++ OK, passed 100 tests.

Productos de polinomios

productoMM :: Monomio -> Monomio -> Monomio

(productoMM m1 m2) es el producto de los monomios m1 y m2. Por ejemplo,

>>> productoMM (M ["x","y"]) (M ["y","z"])
x*y*z

productoMP :: Monomio -> Polinomio -> Polinomio

(productoMP m p) es el producto del monomio m por el polinomio p. Por ejemplo,

>>> productoMP (M ["x","y"]) (P [M ["a"], M ["y","z"]])
a*x*y+x*y*z

producto :: Polinomio -> Polinomio -> Polinomio

(producto p1 p2) es el producto de los polinomios p1 y p2. Por ejemplo,

>>> producto (P [M ["x","y"], M ["z"]]) (P [M ["a"], M ["y","z"]])
a*x*y+a*z+x*y*z+y*z
>>> producto (P [M ["x","y"], M ["z"]]) (P [M ["x"], M ["y","z"]])
x*y+x*y*z+x*z+y*z
>>> producto (P [M ["x"], M ["y"]]) (P [M ["x"], M ["y"]])
x+y

prop_prod_bien_definido :: Polinomio -> Polinomio -> Bool

Comprueba que el producto de polinomios está bien definido.

>>> quickCheck prop_prod_bien_definido
+++ OK, passed 100 tests.

prop_prod_conmutativa :: Polinomio -> Polinomio -> Bool

Comprueba que el producto de polinomios es conmutativo

>>> quickCheck prop_prod_conmutativa
+++ OK, passed 100 tests.

prop_distributiva :: Polinomio -> Polinomio -> Polinomio -> Bool

Comprueba que el producto de polinomios es distributivo respecto de la suma.

>>> quickCheck prop_distributiva
+++ OK, passed 100 tests.

Derivada de polinomios

deriv :: Polinomio -> Variable -> Polinomio

(deriv p x) es la derivada del polinomio p respecto de la variable x; es decir, la lista de monomios de p que contienen la variable x, eliminándola. Por ejemplo,

>>> deriv (P [M ["u"],M ["x"],M ["x","y"],M ["x","z"]]) "x"
1+y+z

prop_deriv_bien_definida :: Polinomio -> Variable -> Bool

Comprueba que la derivada está bien definida.

>>> quickCheck prop_deriv_bien_definida
+++ OK, passed 100 tests.

prop_deriv_deriv :: Polinomio -> Variable -> Bool

Comprueba que la segunda derivada es nula.

>>> quickCheck prop_deriv_deriv
+++ OK, passed 100 tests.

prop_deriv_suma :: Polinomio -> Polinomio -> Variable -> Bool

Comprueba que la derivada de la suma es la suma de las derivadas.

>>> quickCheck prop_deriv_suma
+++ OK, passed 100 tests.

prop_deriv_prod :: Polinomio -> Polinomio -> Variable -> Bool

Comprueba que la regla de la derivada del producto

>>> quickCheck prop_deriv_prod
+++ OK, passed 100 tests.